fix(nodejs-lib-ci): Use `test` script instead of `test:ci:unit`
This means that we will no longer test the transpiled code. We'll run the original TS from now on.

I originally wanted to test the transpiled code for good measure, to catch compilation issues early on, but it's been a pain in the neck to maintain the two test suites in practice and I don't recall a single instance where that actually caught transpilation issues.